:bug: update experiment setup

+++ main.sh
@@ -1,24 +1,40 @@
 #!/usr/bin/env bash
 src="rak-communities-openmp"
-out="/home/resources/Documents/subhajit/$src.log"
-ulimit -c unlimited
+out="$HOME/Logs/$src$1.log"
 ulimit -s unlimited
 printf "" > "$out"
 
 # Download program
-rm -rf $src
-git clone https://github.com/puzzlef/$src
-cd $src
+if [[ "$DOWNLOAD" != "0" ]]; then
+  rm -rf $src
+  git clone https://github.com/puzzlef/$src
+  cd $src
+fi
+
+# Fixed config
+: "${TYPE:=float}"
+: "${MAX_THREADS:=64}"
+: "${REPEAT_METHOD:=5}"
+# Define macros (dont forget to add here)
+DEFINES=(""
+"-DTYPE=$TYPE"
+"-DMAX_THREADS=$MAX_THREADS"
+"-DREPEAT_METHOD=$REPEAT_METHOD"
+)
 
 # Run
-g++ -std=c++17 -O3 -fopenmp -g -rdynamic main.cxx
-stdbuf --output=L ./a.out ~/data/GAP-road.mtx     2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/it-2004.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/webbase-2001.mtx 2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/GAP-twitter.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/GAP-web.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/sk-2005.mtx      2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/GAP-kron.mtx     2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/GAP-urand.mtx    2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/AGATHA_2015.mtx  2>&1 | tee -a "$out"
-stdbuf --output=L ./a.out ~/data/MOLIERE_2016.mtx 2>&1 | tee -a "$out"
+g++ ${DEFINES[*]} -std=c++17 -O3 -fopenmp main.cxx
+# stdbuf --output=L ./a.out ~/Data/web-Stanford.mtx   0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/indochina-2004.mtx  0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/uk-2002.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/arabic-2005.mtx     0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/uk-2005.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/webbase-2001.mtx    0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/it-2004.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/sk-2005.mtx         0 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/com-LiveJournal.mtx 1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/com-Orkut.mtx       1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/asia_osm.mtx        1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/europe_osm.mtx      1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/kmer_A2a.mtx        1 0 2>&1 | tee -a "$out"
+stdbuf --output=L ./a.out ~/Data/kmer_V1r.mtx        1 0 2>&1 | tee -a "$out"
